Revert "feat: add event filters to event log table (#207)"
This reverts commit 6c47c1287bc49af0578b290cc13700818dcd5ff4.
