+++ package.json
@@ -60,7 +60,6 @@
     "react-custom-scrollbars": "^4.2.1",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
-    "react-select": "^3.1.0",
     "react-toastify": "^6.0.5",
     "react-toggle": "^4.1.1",
     "react-virtualized-auto-sizer": "^1.0.2",
